+++ blocks/list/src/index.tsx
@@ -1,79 +1,149 @@
 import { bootstrap, FormattedMessage } from '@appsemble/preact';
 import { Loader, Message } from '@appsemble/preact-components';
+import { type VNode } from 'preact';
 import { useCallback, useEffect, useState } from 'preact/hooks';
 
+import { CollapsibleListComponent } from './components/CollapsibleList/index.js';
 import { ListItem } from './components/ListItem/index.js';
+import styles from './index.module.css';
 import { type Item } from '../block.js';
 
-bootstrap(({ data: blockData, events, parameters: { base }, ready, utils }) => {
-  const [data, setData] = useState<Item[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(false);
+export const renderItems = (items: Item[], spaced?: boolean): VNode => (
+  <ul className={spaced ? 'py-4 px-5' : ''}>
+    {items.map((item, index) => (
+      <li key={item.id ?? index}>
+        <ListItem index={index} item={item} />
+      </li>
+    ))}
+  </ul>
+);
 
-  useEffect(() => {
-    if (blockData != null) {
-      const newData = base == null ? blockData : (blockData as any)[base];
+bootstrap(
+  ({
+    data: blockData,
+    events,
+    parameters: { base, collapsible, groupBy, hideOnNoData = false, title },
+    ready,
+    utils,
+  }) => {
+    const [data, setData] = useState<Item[]>([]);
+    const [groupedData, setGroupedData] = useState<Record<string, Item[]>>({});
+    const [loading, setLoading] = useState(true);
+    const [error, setError] = useState(false);
 
-      if (Array.isArray(newData)) {
-        setData(newData);
-        setLoading(false);
-      }
-    }
-  }, [base, blockData]);
+    useEffect(() => {
+      if (blockData != null) {
+        const newData = base == null ? blockData : (blockData as any)[base];
 
-  const loadData = useCallback(
-    (d: any, err: string): void => {
-      if (err) {
-        setError(true);
-      } else {
-        if (base == null) {
-          setData(d);
-        } else {
-          setData(d[base]);
+        if (Array.isArray(newData)) {
+          setData(newData);
+          setLoading(false);
         }
-        setError(false);
       }
-      setLoading(false);
-    },
-    [base],
-  );
+    }, [base, blockData]);
 
-  useEffect(() => {
-    events.on.data(loadData);
-    ready();
-  }, [events, loadData, ready, utils]);
+    useEffect(() => {
+      if (data != null) {
+        const newGroupedData: Record<string, Item[]> = {};
+        for (const entry of data) {
+          const groupName = entry[groupBy];
+          if (groupName && typeof groupName === 'string') {
+            newGroupedData[groupName] = [...(newGroupedData[groupName] ?? []), entry];
+          }
+        }
+        setGroupedData(newGroupedData);
+      }
+    }, [data, groupBy]);
 
-  if (loading) {
-    return <Loader />;
-  }
-
-  if (error) {
-    return (
-      <Message className="mt-4 mr-6 mb-4 ml-5" color="danger">
-        <span>
-          <FormattedMessage id="error" />
-        </span>
-      </Message>
+    const loadData = useCallback(
+      (d: any, err: string): void => {
+        if (err) {
+          setError(true);
+        } else {
+          if (base == null) {
+            setData(d);
+          } else {
+            setData(d[base]);
+          }
+          setError(false);
+        }
+        setLoading(false);
+      },
+      [base],
     );
-  }
 
-  if (!data?.length) {
-    return (
-      <Message className="mt-4 mr-6 mb-4 ml-5">
-        <span>
-          <FormattedMessage id="noData" />
-        </span>
-      </Message>
+    useEffect(() => {
+      events.on.data(loadData);
+      ready();
+    }, [events, loadData, ready, utils]);
+
+    if (loading) {
+      return <Loader />;
+    }
+
+    if (hideOnNoData && !data?.length) {
+      return null;
+    }
+
+    if (error) {
+      return (
+        <Message className="mt-4 mr-6 mb-4 ml-5" color="danger">
+          <span>
+            <FormattedMessage id="error" />
+          </span>
+        </Message>
+      );
+    }
+
+    if (!data?.length) {
+      return (
+        <Message className="mt-4 mr-6 mb-4 ml-5">
+          <span>
+            <FormattedMessage id="noData" />
+          </span>
+        </Message>
+      );
+    }
+
+    const renderFirstList = (): VNode => (
+      <CollapsibleListComponent
+        index={0}
+        items={data}
+        title={utils.remap(title, blockData) as string}
+      />
     );
-  }
 
-  return (
-    <ul className="py-4 px-5">
-      {data.map((item, index) => (
-        <li key={item.id ?? index}>
-          <ListItem index={index} item={item} />
-        </li>
-      ))}
-    </ul>
-  );
-});
+    return groupBy ? (
+      collapsible ? (
+        <div>
+          {Object.entries(groupedData).length
+            ? Object.entries(groupedData).map(([key, value], index) => (
+                <CollapsibleListComponent index={index} items={value} key={key} title={key} />
+              ))
+            : renderFirstList()}
+        </div>
+      ) : (
+        <div>
+          {Object.entries(groupedData).length ? (
+            Object.entries(groupedData).map(([key, value]) => (
+              <div key={key}>
+                <div className={styles.title}>{key}</div>
+                {renderItems(value)}
+              </div>
+            ))
+          ) : (
+            <>
+              <div className={styles.title}>{title}</div>
+              {renderItems(data)}
+            </>
+          )}
+        </div>
+      )
+    ) : (
+      <>
+        {title && !collapsible ? <div className={styles.title}>{title}</div> : null}
+        {collapsible ? renderFirstList() : renderItems(data, true)}
+      </>
+    );
+  },
+);