+++ packages/app/components/TabsPage/TabContent/index.tsx
@@ -1,5 +1,5 @@
-import { useMeta } from '@appsemble/react-components';
-import { type ComponentPropsWithoutRef, type ReactNode } from 'react';
+import { applyRefs, useMeta } from '@appsemble/react-components';
+import { type ComponentPropsWithoutRef, type MutableRefObject, type ReactNode } from 'react';
 
 import { BlockList } from '../../BlockList/index.js';
 
@@ -10,13 +10,17 @@ interface TabContentProps extends ComponentPropsWithoutRef<typeof BlockList> {
    * This will be set in the document title.
    */
   readonly name: string;
+
+  readonly tabRef: MutableRefObject<unknown>;
 }
 
 /**
  * Render content for a single tab page.
  */
-export function TabContent({ name, ...props }: TabContentProps): ReactNode {
+export function TabContent({ name, tabRef, ...props }: TabContentProps): ReactNode {
   useMeta(name);
+
+  applyRefs({ name }, tabRef);
 
   return <BlockList {...props} />;
 }