+++ packages/node-utils/server/controllers/index.ts
@@ -3,5 +3,6 @@ export * from './appMessages.js';
 export * from './assets.js';
 export * from './resources.js';
 export * from './appMembers.js';
+export * from './appVariables.js';
 export * from './teams.js';
 export * from './oauth2Provider.js';