+++ packages/server/controllers/oauth2Login.ts
@@ -7,7 +7,7 @@ import { type Transaction } from 'sequelize';
 import { EmailAuthorization, OAuthAuthorization, transactional, User } from '../models/index.js';
 import { argv } from '../utils/argv.js';
 import { createJWTResponse } from '../utils/createJWTResponse.js';
-import { type Mailer, type Recipient } from '../utils/email/Mailer.js';
+import { type Mailer } from '../utils/email/Mailer.js';
 import { getAccessToken, getUserInfo } from '../utils/oauth2.js';
 import { githubPreset, gitlabPreset, googlePreset, presets } from '../utils/OAuth2Presets.js';
 
@@ -24,10 +24,20 @@ const processEmailAuthorization = async (
     { UserId: id, email: email.toLowerCase(), key, verified },
     { transaction },
   );
+  const user = await User.findByPk(id, { attributes: ['locale'] });
   if (!verified) {
-    await mailer.sendTemplateEmail({ email, name } as Recipient, 'resend', {
-      url: `${argv.host}/verify?token=${key}`,
-      name: 'The Appsemble Team',
+    await mailer.sendTranslatedEmail({
+      to: {
+        name,
+        email,
+      },
+      emailName: 'resend',
+      ...(user ? { locale: user.locale } : {}),
+      values: {
+        link: (text) => `[${text}](${argv.host}/verify?token=${key})`,
+        name: user ? user.name : 'null',
+        appName: 'null',
+      },
     });
   }
 };