+++ packages/studio/pages/apps/app/index.tsx
@@ -63,19 +63,21 @@ const EditPage = lazy(() => import('./edit/index.js'));
 const GuiEditorPage = lazy(() => import('./GuiEditor/index.js'));
 
 export function AppRoutes(): ReactNode {
-  const { id, lang } = useParams<{ id: string; lang: string }>();
-  const url = `apps/${id}`;
+  const {
+    '*': splat,
+    id,
+    lang,
+    path,
+  } = useParams<{ id: string; lang: string; path?: string; '*': string }>();
   const { organizations } = useUser();
   const {
     data: app,
     error,
     loading,
     setData: setApp,
-  } = useData<App>(
-    Number.isNaN(Number.parseInt(id))
-      ? `/api/apps/path/${id}?language=${lang}`
-      : `/api/apps/${id}?language=${lang}`,
-  );
+  } = useData<App>(`/api/apps/${id}?language=${lang}`);
+
+  const url = `apps/${id}/${path}`;
 
   const { formatMessage } = useIntl();
 
@@ -204,6 +206,15 @@ export function AppRoutes(): ReactNode {
 
   if (!organizations || loading) {
     return <Loader />;
+  }
+
+  if (path === undefined) {
+    /* For compatibility with old app links, see #1490 */
+    return <Navigate replace to={`/${lang}/apps/${id}/${app.path}`} />;
+  }
+
+  if (app && app.path !== path) {
+    return <Navigate replace to={`/${lang}/apps/${id}/${app.path}/${splat}`} />;
   }
 
   return (