+++ packages/utils/api/index.test.ts
@@ -72,6 +72,7 @@ describe('schemas', () => {
     it.each(entries.filter(([, schema]) => schema.type === 'object'))('%s', (path, schema) => {
       // Action descriptions are defined on the type property.
       const descriptionSchema = path.endsWith('ActionDefinition') ? schema.properties.type : schema;
+      const isTabsPageDefinition = path.endsWith('TabsPageDefinition');
       expect(schema).not.toHaveProperty('additionalItems');
       expect(schema).toHaveProperty('additionalProperties');
       expect(schema).not.toHaveProperty('allOf');
@@ -91,7 +92,10 @@ describe('schemas', () => {
       expect(schema).not.toHaveProperty('minItems');
       expect(schema).not.toHaveProperty('minLength');
       expect(schema).not.toHaveProperty('multipleOf');
-      expect(schema).not.toHaveProperty('oneOf');
+      if (!isTabsPageDefinition) {
+        // eslint-disable-next-line vitest/no-conditional-expect
+        expect(schema).not.toHaveProperty('oneOf');
+      }
       expect(schema).not.toHaveProperty('then');
       expect(schema).not.toHaveProperty('uniqueItems');
       if (schema.required && !schema.additionalProperties) {